refactor: 로그인 여부에 따라 메뉴 사이드바 컴포넌트 분리 및 명칭 개선 (#94)
- 사이드바 메뉴 컴포넌트 이름을 명확하게 수정
- 로그인 여부에 따라 메뉴 UI를 컴포넌트 단위로 분리
- 관련된 store 로직은 그대로 유지
